Config API for Kubernetes cluster access information
Introduces a new sysinv config API for retrieving the Kubernetes
cluster access information and security credentials (if configured).

The information may be used to configure a remote Kubernetes client
with the required configuration to use the Kubernetes API with
administrative privileges with either client certificate authentication
or token authentication for the kubernetes-admin service account.

The following information is available for each cluster:
  Kubernetes Cluster Name (kubernetes)
  Kubernetes Release Version
  Cluster API Endpoint URL
  Cluster Root CA Certificate
  Admin Client Certificate
  Admin Client Key
  Admin User Name (kubernetes-admin)
  Admin service account token

Story: 2008630
Task: 41836
Signed-off-by: Matt Peters <matt.peters@windriver.com>
Change-Id: Ib81c7fcc3a577c1209ab3a0dd882552ba3d2b9db